pyproject.toml
[project]
name = "amazon-sp-agent-mvp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.0.0",
    "pandas-stubs==2.3.0.250703",
    "pyyaml>=6.0",
]
